monobj_boss: logical shift for script[0x1A/2] >> 1 in calcBranchFuncGolem
The target shifts script[0x1A/2] (an unsigned short) with srwi (logical), then
compares it signed against script[0x1C/2] via the subfc/adde 33-bit sequence.
The previous unguarded >> promoted the unsigned short to signed int, yielding
srawi plus a clrlwi cleanup and reordered loads. Casting the shift operand to
unsigned int (logical srwi) and back to int (signed compare preserved) matches
the load order and shift. Behavior is identical for the 0..65535 range.
calcBranchFuncGolem 88.31% -> 93.75% (only a trailing bool-widening clrlwi left).

+++ src/monobj_boss.cpp
@@ -446,7 +446,9 @@ int CGMonObj::calcBranchFuncGolem(int)
 	}
 
 	unsigned short* script = reinterpret_cast<unsigned short*>(object->m_scriptHandle);
-	return static_cast<unsigned int>(__cntlzw(static_cast<unsigned int>(script[0x1C / 2] >= (script[0x1A / 2] >> 1)))) >> 5;
+	return static_cast<unsigned int>(
+	           __cntlzw(script[0x1C / 2] >= static_cast<int>(static_cast<unsigned int>(script[0x1A / 2]) >> 1))) >>
+	       5;
 }
 
 /*
